# Conflicts:
#	src/main/java/org/example/Main.java
#	src/main/java/persistence/PlayerDaoImpl.java
#	src/main/java/services/PlayerService.java

+++ src/main/java/model/entity/Player.java
@@ -26,10 +26,10 @@ public class Player {
 	private long idArea;
 
 	@Column(name = "X_AXIS")
-	private long x_axis=0;
+	private int x_axis=0;
 
 	@Column(name = "Y_AXIS")
-	private long y_axis=0;
+	private int y_axis=0;
 
 	@OneToOne(cascade = CascadeType.ALL)
 	@JoinColumn(name = "id_inventory", referencedColumnName = "id")
@@ -96,19 +96,19 @@ public class Player {
 		this.skills = skills;
 	}
 
-	public long getY_axis() {
+	public int getY_axis() {
 		return y_axis;
 	}
 
-	public void setY_axis(long y_axis) {
+	public void setY_axis(int y_axis) {
 		this.y_axis = y_axis;
 	}
 
-	public long getX_axis() {
+	public int getX_axis() {
 		return x_axis;
 	}
 
-	public void setX_axis(long x_axis) {
+	public void setX_axis(int x_axis) {
 		this.x_axis = x_axis;
 	}
 }
